pyproject.toml
[project]
name = "bookshelf"
version = "0.1.0"
description = "Bookshelf API with Pinecone integration"
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
    "fastapi>=0.116.1",
    "mammoth>=1.10.0",
    "numpy>=2.3.2",
    "openai>=1.99.6",
    "pinecone>=7.3.0",
    "pymupdf>=1.26.3",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.1.0",
    "tiktoken>=0.11.0",
    "uvicorn[standard]>=0.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

